+++ src/ultimate_notion/__init__.py
@@ -49,8 +49,8 @@ from ultimate_notion.core import get_active_session
 from ultimate_notion.database import Database
 from ultimate_notion.emoji import Emoji
 from ultimate_notion.file import FileInfo
-from ultimate_notion.obj_api.enums import AggFunc, BGColor, CodeLang, Color, NumberFormat, VState
-from ultimate_notion.option import Option, OptionNS
+from ultimate_notion.obj_api.enums import AggFunc, BGColor, CodeLang, Color, NumberFormat, OptionGroupType, VState
+from ultimate_notion.option import Option, OptionGroup, OptionNS
 from ultimate_notion.page import Page
 from ultimate_notion.query import Condition, prop
 from ultimate_notion.rich_text import join, math, mention, text
@@ -90,6 +90,8 @@ __all__ = [
     'NumberFormat',
     'NumberedItem',
     'Option',
+    'OptionGroup',
+    'OptionGroupType',
     'OptionNS',
     'Page',
     'Paragraph',